+++ TSEmployeeProject/TSEmployeeProject/Models/Position.cs
@@ -9,7 +9,7 @@ namespace TSEmployeeProject.Models
     public partial class Position
     {
         [JsonProperty("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [JsonProperty("name")]
         public string Name { get; set; }
@@ -18,6 +18,6 @@ namespace TSEmployeeProject.Models
         public string Level { get; set; }
 
         [JsonProperty("sort")]
-        public long Sort { get; set; }
+        public int Sort { get; set; }
     }
 }
